Decompose homepage into section components
